% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{fastlm}
\alias{fastlm}
\alias{fastlm-package}
\title{fastlm}
\usage{
fastlm(x, y)
}
\arguments{
\item{x}{numeric the independent variable}

\item{y}{numeric the dependent variable}
}
\value{
list with three parameters: slope, intercept, and RSql
}
\description{
fastlm

Calculate Slope, Intercept and Rsq
}

